backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
pillow>=10.0.0
pillow-heif>=0.13.0
opencv-python-headless>=4.8.0
piexif>=1.1.3
exifread>=3.0.0
imagehash>=4.3.1
xxhash>=3.4.0
gpxpy>=1.6.0
send2trash>=1.8.0
python-magic-bin>=0.4.14
torch>=2.1.0
open-clip-torch>=2.23.0
numpy>=1.24.0
websockets>=12.0
pydantic>=2.0.0
